New 'configure' setup.py command
This step is taking care of running all the necessary checks that configure the
pyprofit extension (which we were otherwise doing every time we run the
setup.py script). The configure command is then run before the build_ext
command, and thus ensure that the extension is configured before it's built,
while at the same time avoiding running all the configuration steps when not
required.

Signed-off-by: Rodrigo Tobar <rtobar@icrar.org>