Improving libprofit lookup mechanism
Default system-level directories are now correctly calculated, which
were previously not (so installations under /usr or /usr/local were not
being picked up). Also, user-provided locations now take precedence,
which again wasn't the case before.

More improvements could be further incorporated in the future (e.g.,
always check for the library that corresponds *exactly* to the header
found), but that requires a bit more of work and careful testing, and
therefore time.

Signed-off-by: Rodrigo Tobar <rtobar@icrar.org>